requirements.txt
accelerate==0.34.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
av==13.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
clip==1.0
coloredlogs==15.0.1
conda-pack==0.8.1
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dashscope==1.24.9
diffusers==0.31.0
dominate==2.9.1
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.19.1
flash_attn==2.5.8
Flask==3.1.2
Flask-SocketIO==5.5.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
MarkupSafe==2.1.5
matplotlib==3.10.7
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.3
ninja==1.13.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.19.1
onnx-ir==0.1.12
onnxconverter-common==1.16.0
onnxruntime==1.23.2
onnxscript==0.5.6
open_clip_torch==3.2.0
opencv-python==4.11.0.86
packaging==25.0
peft==0.13.0
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pycocotools==2.0.10
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.14.3
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
self_forcing==0.0.1
self_forcing==0.0.1
sentencepiece==0.2.1
sentry-sdk==2.43.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
sympy==1.13.1
tifffile==2025.5.10
timm==1.0.21
tokenizers==0.19.1
torch==2.5.1+cu121
torchao==0.14.1
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.44.2
triton==3.1.0
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.22.3
wcwidth==0.2.14
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0
